+++ bank-feeds/Codat/BankFeeds/Models/Operations/ListSourceAccountsResponse.cs
@@ -9,9 +9,9 @@
 #nullable enable
 namespace Codat.BankFeeds.Models.Operations
 {
+    using Codat.BankFeeds.Models.Operations;
     using Codat.BankFeeds.Models.Shared;
     using Codat.BankFeeds.Utils;
-    using System.Collections.Generic;
     using System.Net.Http;
     using System;
     
@@ -36,6 +36,6 @@ namespace Codat.BankFeeds.Models.Operations
         /// <summary>
         /// Success
         /// </summary>
-        public List<SourceAccountV2>? SourceAccounts { get; set; }
+        public ListSourceAccountsResponseBody? OneOf { get; set; }
     }
 }